requirements.txt
absl-py==0.7.0
astor==0.8.1
astunparse==1.6.3
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dill==0.3.2
future==0.18.2
gast==0.2.2
google-api-core==1.22.1
google-api-python-client==1.10.1
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
jsonpickle==1.4.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
nltk==3.5
numpy==1.18.5
oauth2client==3.0.0
oauthlib==3.1.0
opt-einsum==3.3.0
promise==2.3
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyTelegramBotAPI==3.7.3
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
SQLAlchemy~=1.3.18
tensorboard==2.0.2
tensorboard-plugin-wit==1.7.0
tensorflow~=2.3.0
tensorflow-datasets==3.2.1
tensorflow-estimator==2.0.1
tensorflow-metadata==0.23.0
termcolor==1.1.0
textblob==0.15.3
threadpoolctl==2.1.0
tqdm==4.48.2
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
